refactor: 순환 참조 제거
